Splitting matplotlib code from the rest of plotting stuff
