+++ pandas/plotting/_matplotlib/compat.py
