+++ pandas/plotting/_matplotlib/style.py
@@ -1,7 +1,7 @@
 # being a bit too dynamic
-from contextlib import contextmanager
 import warnings
 
+import matplotlib.pyplot as plt
 import numpy as np
 
 from pandas.core.dtypes.common import is_list_like
@@ -9,8 +9,6 @@ from pandas.core.dtypes.common import is_list_like
 
 def _get_standard_colors(num_colors=None, colormap=None, color_type='default',
                          color=None):
-    import matplotlib.pyplot as plt
-
     if color is None and colormap is not None:
         if isinstance(colormap, str):
             import matplotlib.cm as cm
@@ -91,72 +89,3 @@ def _get_standard_colors(num_colors=None, colormap=None, color_type='default',
         colors += colors[:mod]
 
     return colors
-
-
-class _Options(dict):
-    """
-    Stores pandas plotting options.
-    Allows for parameter aliasing so you can just use parameter names that are
-    the same as the plot function parameters, but is stored in a canonical
-    format that makes it easy to breakdown into groups later
-    """
-
-    # alias so the names are same as plotting method parameter names
-    _ALIASES = {'x_compat': 'xaxis.compat'}
-    _DEFAULT_KEYS = ['xaxis.compat']
-
-    def __init__(self, deprecated=False):
-        self._deprecated = deprecated
-        # self['xaxis.compat'] = False
-        super().__setitem__('xaxis.compat', False)
-
-    def __getitem__(self, key):
-        key = self._get_canonical_key(key)
-        if key not in self:
-            raise ValueError(
-                '{key} is not a valid pandas plotting option'.format(key=key))
-        return super().__getitem__(key)
-
-    def __setitem__(self, key, value):
-        key = self._get_canonical_key(key)
-        return super().__setitem__(key, value)
-
-    def __delitem__(self, key):
-        key = self._get_canonical_key(key)
-        if key in self._DEFAULT_KEYS:
-            raise ValueError(
-                'Cannot remove default parameter {key}'.format(key=key))
-        return super().__delitem__(key)
-
-    def __contains__(self, key):
-        key = self._get_canonical_key(key)
-        return super().__contains__(key)
-
-    def reset(self):
-        """
-        Reset the option store to its initial state
-
-        Returns
-        -------
-        None
-        """
-        self.__init__()
-
-    def _get_canonical_key(self, key):
-        return self._ALIASES.get(key, key)
-
-    @contextmanager
-    def use(self, key, value):
-        """
-        Temporarily set a parameter value using the with statement.
-        Aliasing allowed.
-        """
-        old_value = self[key]
-        try:
-            self[key] = value
-            yield self
-        finally:
-            self[key] = old_value
-
-
-plot_params = _Options()
