+++ pandas/tests/plotting/test_datetimelike.py
@@ -161,7 +161,7 @@ class TestTSPlot(TestPlotBase):
             _check_plot_works(ser.plot, ax=ax)
 
     def test_get_datevalue(self):
-        from pandas.plotting._converter import get_datevalue
+        from pandas.plotting._matplotlib.converter import get_datevalue
         assert get_datevalue(None, 'D') is None
         assert get_datevalue(1987, 'A') == 1987
         assert (get_datevalue(Period(1987, 'A'), 'M') ==
@@ -285,7 +285,7 @@ class TestTSPlot(TestPlotBase):
 
     @pytest.mark.slow
     def test_uhf(self):
-        import pandas.plotting._converter as conv
+        import pandas.plotting._matplotlib.converter as conv
         idx = date_range('2012-6-22 21:59:51.960928', freq='L', periods=500)
         df = DataFrame(np.random.randn(len(idx), 2), index=idx)
 
@@ -416,7 +416,7 @@ class TestTSPlot(TestPlotBase):
             _test(ax)
 
     def test_get_finder(self):
-        import pandas.plotting._converter as conv
+        import pandas.plotting._matplotlib.converter as conv
 
         assert conv.get_finder('B') == conv._daily_finder
         assert conv.get_finder('D') == conv._daily_finder
